+++ src/app/dev/DevToys.Tools/Helpers/ListCompareHelper.cs
@@ -1,4 +1,5 @@
-﻿using DevToys.Tools.Models;
+﻿using DevToys.Tools.Helpers.Core;
+using DevToys.Tools.Models;
 using Markdig.Helpers;
 using Microsoft.Extensions.Logging;
 
@@ -88,23 +89,4 @@ internal static partial class ListCompareHelper
     }
 }
 
-internal class ReadOnlyMemoryEqualityComparer : IEqualityComparer<ReadOnlyMemory<char>>
-{
-    private readonly StringComparison _comparisonType;
-
-    public ReadOnlyMemoryEqualityComparer(bool caseSensitive)
     {
-        _comparisonType = caseSensitive ? StringComparison.Ordinal : StringComparison.OrdinalIgnoreCase;
-    }
-
-
-    public bool Equals(ReadOnlyMemory<char> x, ReadOnlyMemory<char> y)
-    {
-        return x.Span.Equals(y.Span, _comparisonType);
-    }
-
-    public int GetHashCode(ReadOnlyMemory<char> obj)
-    {
-        return obj.ToString().GetHashCode();
-    }
-}
